finnhub/requirements.txt
websocket-client == 1.7.0
avro == 1.11.3
kafka-python == 2.0.2
pyyaml == 5.3.1